requirements.txt
# Core Framework
crewai==0.28.0
crewai-tools==0.2.0
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-openai==0.0.5

# LLM Providers
anthropic==0.18.0
openai==1.12.0

# Document Parsing
pymupdf4llm==0.0.5
PyMuPDF>=1.23.0            # Provides fitz module for PDF fallback
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.2
beautifulsoup4==4.12.3

# Image Processing
Pillow==10.2.0
replicate==0.25.0

# Platform Publishers
praw==7.7.1              # Reddit
tweepy==4.14.0            # X/Twitter

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pytz==2024.1
tzlocal>=5.0               # System timezone detection
requests==2.31.0
aiohttp==3.9.1
celery==5.3.4
redis==5.0.1

# Web Framework (for API service)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
